Moved TimeDiffFormatter to the common module.
