+++ vpn-store/src/main/java/com/duckduckgo/mobile/android/vpn/dao/VpnRestartsDao.kt
@@ -17,7 +17,7 @@
 package com.duckduckgo.mobile.android.vpn.dao
 
 import androidx.room.*
-import com.duckduckgo.mobile.android.vpn.store.DatabaseDateFormatter
+import com.duckduckgo.app.global.formatters.time.DatabaseDateFormatter
 
 @Entity(tableName = "vpn_phoenix")
 data class VpnPhoenixEntity(
