+++ vpn-store/src/main/java/com/duckduckgo/mobile/android/vpn/dao/VpnRunningStatsDao.kt
@@ -18,7 +18,7 @@ package com.duckduckgo.mobile.android.vpn.dao
 
 import androidx.room.*
 import com.duckduckgo.mobile.android.vpn.model.VpnRunningStats
-import com.duckduckgo.mobile.android.vpn.store.DatabaseDateFormatter
+import com.duckduckgo.app.global.formatters.time.DatabaseDateFormatter
 import kotlinx.coroutines.flow.Flow
 
 @Dao
