+++ vpn-store/src/main/java/com/duckduckgo/mobile/android/vpn/model/VpnDatabaseModels.kt
@@ -19,7 +19,7 @@
 package com.duckduckgo.mobile.android.vpn.model
 
 import androidx.room.*
-import com.duckduckgo.mobile.android.vpn.store.DatabaseDateFormatter
+import com.duckduckgo.app.global.formatters.time.DatabaseDateFormatter
 import com.duckduckgo.mobile.android.vpn.trackers.AppTrackerEntity
 
 @Entity(
