+++ vpn/src/main/java/com/duckduckgo/mobile/android/vpn/stats/AppTrackerBlockingStatsRepository.kt
@@ -19,7 +19,7 @@ package com.duckduckgo.mobile.android.vpn.stats
 import androidx.annotation.WorkerThread
 import com.duckduckgo.mobile.android.vpn.dao.VpnPhoenixEntity
 import com.duckduckgo.mobile.android.vpn.model.*
-import com.duckduckgo.mobile.android.vpn.store.DatabaseDateFormatter
+import com.duckduckgo.app.global.formatters.time.DatabaseDateFormatter
 import com.duckduckgo.mobile.android.vpn.store.VpnDatabase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.*
